Improved Modularity of Routes
Now each route is much more ndependent and simpler to view and change.
This also allows for easier testing of specific features. A constants
file was also added for the database url to make it changeable in one
place. The game.pug file was slightly modified to allow for input.